Moved 'pso_example' function to file 'examples.rs'
