+++ src/main.rs
@@ -3,6 +3,7 @@ mod ga;
 mod pso;
 mod ff;
 mod examples;
+mod test_functions;
 
 fn main() {
 	examples::ga_example()
